requirements.txt

anyio==3.6.2 

fastapi==0.79.0
fastapi==0.88.0

h11==0.14.0
idna==3.4 
orjson==3.4.1
pydantic==1.9.0
pydantic==1.10.2

pyyaml==6.0
sniffio==1.3.0
starlette==0.22.0
uvicorn==0.18.2
uvicorn==0.20.0
uvloop==0.14.0
